+++ .standard.yml
@@ -1,6 +1,6 @@
 # I really just have issues with the automatic "semantic blocks"
 
-ruby_version: 2.4
+ruby_version: 2.7
 
 format: progress
