+++ spec/spec_helper.rb
@@ -27,3 +27,11 @@ class MockSyslog
     @output << [severity, message]
   end
 end
+
+RSpec.configure do |config|
+  config.warnings = true
+  config.disable_monkey_patching!
+  config.default_formatter = "doc" if config.files_to_run.one?
+  config.order = :random
+  Kernel.srand config.seed
+end
